Migrate TestFailReporter to JUnit 5 TestFailExtension
Extracts the TestFailReporter inner class from FileBufferFunctions
into a standalone TestFailExtension class that implements
JUnit 5's TestWatcher interface.

Updates FileBufferFunctions to use @RegisterExtension with the
new extension.
